# perf(cli): fast path common argument patterns