--- a/cli/mainrt.rs
+++ b/cli/mainrt.rs
@@ -68,10 +68,9 @@ fn unwrap_or_exit<T>(result: Result<T, AnyError>) -> T {
 }
 
 fn main() {
-  let args: Vec<String> = env::args().collect();
+  let args: Vec<_> = env::args_os().collect();
   let current_exe_path = current_exe().unwrap();
-  let standalone =
-    standalone::extract_standalone(&current_exe_path, args.clone());
+  let standalone = standalone::extract_standalone(&current_exe_path, args);
   let future = async move {
     match standalone {
       Ok(Some(future)) => {