```

15:20 $ hyperfine -S none --warmup 25 '/tmp/deno run /tmp/empty.js' 'target/release/deno run /tmp/empty.js'
Benchmark 1: /tmp/deno run /tmp/empty.js
  Time (mean ± σ):      17.6 ms ±   5.3 ms    [User: 11.4 ms, System: 4.2 ms]
  Range (min … max):    15.1 ms …  85.0 ms    169 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 2: target/release/deno run /tmp/empty.js
  Time (mean ± σ):      16.4 ms ±   0.7 ms    [User: 10.8 ms, System: 4.0 ms]
  Range (min … max):    14.6 ms …  20.1 ms    193 runs
 
Summary
  'target/release/deno run /tmp/empty.js' ran
    1.07 ± 0.33 times faster than '/tmp/deno run /tmp/empty.js'
```